% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleData}
\alias{exampleData}
\title{Example dataset with 218 individuals.}
\source{
Cerón-Rojas, J. Jesús ; Crossa, José; Arief, Vivi N.; Kaye Basford;
        Rutkoski, Jessica; Jarquín, Diego ; Alvarado, Gregorio; Beyene, Yoseph;
        Semagn, Kassa ; DeLacy, Ian, 2015-06-04, "Application of a Genomics
        Selection Index to Real and Simulated Data", \url{http://hdl.handle.net/11529/10199} V10
}
\usage{
exampleData()
}
\value{
Core Hunter data of class \code{chdata}
}
\description{
Includes a precomputed distance matrix read from \code{"extdata/distances.csv"},
genotypes read from \code{"extdata/genotypes.csv"} and phenotypes read from
\code{"extdata/phenotypes.csv"}.
}
\details{
The distance matrix is computed from the genotypes (Modified Rogers' distance).
Data was taken from the CIMMYT Research Data Repository (Study Global ID
hdl:11529/10199; real data set 5, cycle 0).
}

