% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getAlleleFrequencies}
\alias{getAlleleFrequencies}
\title{Get Allele frequency matrix.}
\usage{
getAlleleFrequencies(data)
}
\arguments{
\item{data}{Core Hunter data containing genotypes}
}
\value{
allele frequency matrix
}
\description{
Get Allele frequency matrix.
}
