% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{setRange}
\alias{setRange}
\title{Set the normalization range of the given objective.}
\usage{
setRange(obj, range)
}
\arguments{
\item{obj}{Core Hunter objective of class \code{chobj}.}

\item{range}{Normalization range [l,u].
See argument \code{range} of \code{\link{objective}} for details.}
}
\value{
Objective including normalization range.
}
\description{
See argument \code{range} of \code{\link{objective}} for details.
}
\seealso{
\code{\link{objective}}
}

