% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{wrapData}
\alias{wrapData}
\title{Wrap distances, genotypes or phenotypes in Core Hunter data.}
\usage{
wrapData(data)
}
\arguments{
\item{data}{of class \code{chgeno}, \code{chpheno} or \code{chdist}}
}
\value{
Core Hunter data of class \code{chdata}
}
\description{
If the given data does not match any of these three classes
it is returned unchanged.
}
